requirements.txt
ddgs
langchain-community
sounddevice
scipy
google-generativeai
groq
soundfile
python-dotenv
